Fix assignment data type

+++ app/src/main/java/trianglz/models/Subcategory.java
@@ -28,7 +28,7 @@ public class Subcategory {
     @SerializedName("assignments_total")
     public double assignmentsTotal;
     @SerializedName("assignments_grade")
-    public int assignmentsGrade;
+    public double assignmentsGrade;
     @SerializedName("assignments")
     public ArrayList<Assignments> assignments;
     @SerializedName("grade_items_total")
